+++ src/main/java/com/ceiba/lendings/dominio/entidades/Payment.java
@@ -9,7 +9,7 @@ public class Payment {
     private static final String PAYMENT_DATE_REQUIRED = "La fecha del pago es requerida";
     private static final String PAYMENT_LENDING_REQUIRED = "El prestamo del pago es requerido";
 
-    private Integer id;
+    private Long id;
     private Double payment_value;
     private Date payment_date;
     private Lending lending_id;
@@ -17,7 +17,7 @@ public class Payment {
     public Payment() {
     }
 
-    public Payment(Integer id, Double payment_value, Date payment_date, Lending lending_id) {
+    public Payment(Long id, Double payment_value, Date payment_date, Lending lending_id) {
         PaymentValidator.validateField(payment_date,PAYMENT_DATE_REQUIRED);
         PaymentValidator.validateField(payment_value,PAYMENT_VALUE_REQUIRED);
         PaymentValidator.validateField(lending_id,PAYMENT_LENDING_REQUIRED);
@@ -28,11 +28,11 @@ public class Payment {
         this.lending_id=lending_id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
